small: fix assertion during allocation slab allocation
When `slab_get` is called from `region_alloc` or `ibuf_alloc` first
of all we try to calculate order of appropriate slab. But there is
no check that requested size with slab meta is <= UINT32_MAX, that
leads to assertion failer in `slab_order` function. Add appropriate
check to `slab_get` function, now we return NULL for such big size.