msgpack: add stubs for compressed data support
This commit add stub functions for compressed data
support. Add new MP_EXT type MP_COMPRESSION. Despite
the fact that this type is added to open source version
all functions for working with it are implemented only
in enterprise version of tarantool. In open source version
add stubs for appropriate functions.

Part of #2695