lua: add error handling for lua trigger `push_event` function
Lua trigger `push_event` function is used to pushes th event
data to Lua stack as arguments of the Lua trigger. Previosly
this function never fails, so we don't handle it return code.
But after tuple compression will be implemented, allocation
failure can occur in this function and we must correctly handle
error in this case.

Part of #2695