Implement stubs for tuple compression
This commit add stub functions for tuple compression. Now after new
tuple allocation, we check if space format contains compression for
some fields and if it so, we compress new allocated tuple. When we
return tuple from space to user (from box_select or using C API), we
check if tuple is compressed and if it is so decompress tuple.
In opensource version user unable to set compression for tuple fields,
so all stub functions are unreachable.

Closes #2695