cmake: add `coll` library as a target link library for `tuple` library
Library `tuple` depends on the `coll` library, since `key_def.c` file
from the `tuple` library use functions from the `coll` library. This
usually does not cause problems, since the `coll` library links with
tarantool and resolves symbols in `tuple` library. For tests it is not
a problem, because functions from the `tuple` library, which depends
from the `coll` library are `static inline` and they they are not used
in tests. But adding not `static inline` function in `tuple` library
which depends on `coll` library leads to unresolved symbols in tests.
Fix this explicitly adding `coll` library as a target link library
for `tuple` library.

NO_CHANGELOG=cmake
NO_DOC=cmake